# be/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Hackazon API",
    "description": "Compact reference for core ShopSmart endpoints.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "paths": {
    "/api/v1/user/{id}": {
      "get": {
        "summary": "Fetch user profile",
        "description": "Returns user details without enforcing authentication. Intended for lab exploration.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "User profile returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "summary": "Fetch order by id",
        "description": "Returns order information for any ID supplied.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": 123
          }
        ],
        "responses": {
          "200": {
            "description": "Order information returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          }
        }
      }
    },
    "/api/v1/config": {
      "get": {
        "summary": "Retrieve platform configuration",
        "description": "Returns public configuration details. Supplying a particular admin header reveals extended settings.",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Optional elevated header documented in internal runbooks (value not published)."
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register a new user",
        "description": "Creates a standard user account (role defaults to user).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "User already exists"
          }
        }
      }
    },
    "/api/v1/user/profile": {
      "get": {
        "summary": "View current user's profile",
        "description": "Returns profile details for the authenticated session.",
        "responses": {
          "200": {
            "description": "Profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      },
      "put": {
        "summary": "Update profile",
        "description": "Allows modifying profile fields for the current user. The backend validates very little, so additional JSON properties may be accepted (watch the request body).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "username": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/track": {
      "get": {
        "summary": "Capture affiliate redirect",
        "description": "Records partner activity and then redirects to the requested destination.",
        "parameters": [
          {
            "name": "partner_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifier assigned to the partner generating the referral."
          },
          {
            "name": "redirect",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Encoded URL where the visitor should be sent after tracking."
          }
        ],
        "responses": {
          "302": {
            "description": "User is redirected after tracking."
          },
          "200": {
            "description": "Tracking acknowledgement (used in some lab scenarios)."
          },
          "400": {
            "description": "Missing parameters"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "role": {
            "type": "string",
            "example": "admin"
          },
          "displayName": {
            "type": "string",
            "example": "Administrator"
          },
          "secretNote": {
            "type": "string",
            "example": "PCTFS{user_profile_flag}"
          }
        }
      }
    }
  }
}